Update README.md
Added
Symbol List: Examples of symbols would be like elements on the Periodic Table. Au, O, H, Uuu, Uub, etc. Another example would be like 69420. The Symbol List is the list of all Symbols.
Word List: Examples of words would be like words: Apple, Going, Humming; like names: Andrew, Isaac, Einstein, James, Charles, Newton, Hunter, Jacob, Goliath; or like numbers: 42, 3.14159265, 2.718281828, or 4206942069420. The Word List is the list of all Words.
The goal of the program is that for every Word, it will find every combination of Symbols that can construct the word. If the word cannot be constructed it will skip the word and move onto the next word. The program will then create an Output List that contains every word that can be constructed and the symbols required to construct it. If a word can be spelled through multiple combinations of symbols it will list them as separate entries in the list.
This program will be written in C++.